[Truffle] Remove redundant cast.

+++ truffle/src/main/java/org/jruby/truffle/core/array/ArrayNodes.java
@@ -931,7 +931,7 @@ public abstract class ArrayNodes {
             }
             final Object result = toIntNode.executeIntOrLong(frame, indexObject);
             if (result instanceof Integer) {
-                return (long) (int) result;
+                return (int) result;
             } else {
                 return (long) result;
             }
